requirements.txt
streamlit
python-dotenv
langchain
langchain_community
langchain_groq
duckduckgo-search
wikipedia-api
arxiv
